chunk area and arrow size update
The results table now shows the area (in pixels) of each chunk.

Additionally for arrows, you can change the size of the arrowhead and width of the arrow line.

Label font size option has been simplified and combined for both overlays and complete image output and moved to the Drawing tab (previously the Color tab).